-caricamento progetto DB

+++ bb-project.DB/bb-project.DB/Tables/tbl_Workout.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[tbl_Workout]
 (
-	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
+	[Id] BIGINT NOT NULL PRIMARY KEY IDENTITY, 
     [Name] NVARCHAR(50) NOT NULL, 
     [fk_WorkoutPlanId] BIGINT NOT NULL
 )
